+++ gfx/layers/LayerTreeInvalidation.cpp
@@ -27,7 +27,6 @@
 #include "mozilla/layers/ImageHost.h"
 #include "mozilla/layers/LayerManagerComposite.h"
 #include "TreeTraversal.h"  // for ForEachNode
-#include "LayersLogging.h"
 
 // LayerTreeInvalidation debugging
 #define LTI_DEBUG 0
