+++ gfx/layers/client/TiledContentClient.cpp
@@ -27,7 +27,6 @@
 #include "nsISupportsImpl.h"      // for gfxContext::AddRef, etc
 #include "nsExpirationTracker.h"  // for nsExpirationTracker
 #include "nsMathUtils.h"          // for NS_lroundf
-#include "LayersLogging.h"
 #include "UnitTransforms.h"  // for TransformTo
 #include "mozilla/StaticPrefs_apz.h"
 #include "mozilla/StaticPrefs_layers.h"
