+++ gfx/layers/mlgpu/ContainerLayerMLGPU.cpp
@@ -6,7 +6,6 @@
 
 #include "ContainerLayerMLGPU.h"
 #include "mozilla/StaticPrefs_layers.h"
-#include "LayersLogging.h"
 #include "LayerManagerMLGPU.h"
 #include "MLGDevice.h"
 #include "mozilla/gfx/Rect.h"
