+++ gfx/layers/mlgpu/RenderViewMLGPU.cpp
@@ -9,7 +9,6 @@
 #include "FrameBuilder.h"
 #include "mozilla/StaticPrefs_layers.h"
 #include "LayersHelpers.h"
-#include "LayersLogging.h"
 #include "MLGDevice.h"
 #include "RenderPassMLGPU.h"
 #include "ShaderDefinitionsMLGPU.h"
