+++ gfx/layers/moz.build
@@ -35,7 +35,6 @@ EXPORTS += [
     'ipc/ThreadSafeRefcountingWithMainThreadDestruction.h',
     'Layers.h',
     'LayerScope.h',
-    'LayersLogging.h',
     'LayerSorter.h',
     'LayersTypes.h',
     'LayerTreeInvalidation.h',
@@ -497,7 +496,6 @@ UNIFIED_SOURCES += [
     'Layers.cpp',
     'LayerScope.cpp',
     'LayersHelpers.cpp',
-    'LayersLogging.cpp',
     'LayerSorter.cpp',
     'LayersTypes.cpp',
     'LayerTreeInvalidation.cpp',
